+++ packages/styles/src/packages/atomic/input/rarui-input.types.ts
@@ -1,0 +1,5 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { styles } from "./raui-input.css";
+
+export type InputVariants = RecipeVariants<typeof styles.container> &
+  RecipeVariants<typeof styles.input>;